% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_gwas}
\alias{read_gwas}
\title{Read a GWAS results file into a data frame.}
\usage{
read_gwas(input, missing = c("NA", "N/A", "null", "."), verbose = TRUE)
}
\arguments{
\item{input}{Path to a file containing GWAS summary statistics. If multiple
paths are specified all files will be read in and combined into a single
\code{data.frame}.}

\item{missing}{Vector of characters that represent missing value codes. By
default the following strings are interpreted as \code{NA}: \code{""},
\code{"."}, \code{"NA"}, \code{"N/A"}, and \code{"null"}.}

\item{verbose}{Provide description of processing steps}
}
\description{
Read a GWAS results file into a data frame.
}
