% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.r
\name{set_chromosomes}
\alias{set_chromosomes}
\title{Convert chromosomes column to a propery ordered factor}
\usage{
set_chromosomes(x, style)
}
\description{
Convert chromosomes column to a propery ordered factor
}
