% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write_gwas_csv}
\alias{write_gwas_csv}
\title{Write a GWAS results data frame to a csv file.}
\usage{
write_gwas_csv(x, path, na = "NA", append = FALSE, col_names = !append)
}
\arguments{
\item{x}{A data frame to write to disk}

\item{path}{Path or connection to write to.}

\item{na}{String used for missing values. Defaults to NA. Missing values
will never be quoted; strings with the same value as \code{na} will
always be quoted.}

\item{append}{If \code{FALSE}, will overwrite existing file. If \code{TRUE},
will append to existing file. In both cases, if file does not exist a new
file is created.}

\item{col_names}{Write columns names at the top of the file?}
}
\description{
Write a GWAS results data frame to a csv file.
}
