% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_gwas}
\alias{read_gwas}
\title{Read a GWAS results file into a data frame.}
\usage{
read_gwas(file, verbose = TRUE)
}
\arguments{
\item{verbose}{ Be chatty and report timings? }

\item{delim}{Symbolic name for single character used to separate fields}
}
\description{
Read a GWAS results file into a data frame.
}
\section{Delimiters}{


The \code{delim} argument accepts any of the following values:

\enumerate{
 \item \code{"tab"}
 \item \code{"comma"}
 \item \code{"space"}
}
}

