% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_gwas}
\alias{read_gwas}
\title{Read a GWAS results file into a data frame.}
\usage{
read_gwas(input, sep = "auto", missing = c("NA", "N/A", "null", "."),
  chromosome_style = "ucsc", preprocess = NULL, nrows = -1L,
  header = TRUE, col.names = NULL, verbose = TRUE)
}
\arguments{
\item{input}{Path to a file containing GWAS summary statistics. If multiple
paths are specified all files will be read in and combined into a single
\code{data.frame}.}

\item{sep}{ The separator between columns. Defaults to the first character in the set [\code{,\t |;:}] that exists on line \code{autostart} outside quoted (\code{""}) regions, and separates the rows above \code{autostart} into a consistent number of fields, too. }

\item{missing}{Vector of characters that represent missing value codes. By
default the following strings are interpreted as \code{NA}: \code{""},
\code{"."}, \code{"NA"}, \code{"N/A"}, and \code{"null"}.}

\item{chromosome_style}{Convert chromosomes to ordered factors with labels based on the specified style (default is \code{"ucsc"}; see below for a comparison of the different styles). Set to \code{NULL} to leave chromosomes unchanged.}

\item{preprocess}{a shell command that preprocesses the file; see below for
more details}

\item{nrows}{ The number of rows to read, by default -1 means all. Unlike \code{read.table}, it doesn't help speed to set this to the number of rows in the file (or an estimate), since the number of rows is automatically determined and is already fast. Only set \code{nrows} if you require the first 10 rows, for example. `nrows=0` is a special case that just returns the column names and types; e.g., a dry run for a large file or to quickly check format consistency of a set of files before starting to read any. }

\item{header}{ Does the first data line contain column names? Defaults according to whether every non-empty field on the first data line is type character. If so, or TRUE is supplied, any empty column names are given a default name. }

\item{col.names}{ A vector of optional names for the variables (columns). The default is to use the header column if present or detected, or if not "V" followed by the column number. }

\item{verbose}{Provide description of processing steps}
}
\description{
Read a GWAS results file into a data frame.
}
\section{Chromosome Styles}{


We use the \emph{Homo sapiens} chromosome styles defined in Bioconductor's
\href{https://bioconductor.org/packages/GenomeInfoDb}{\code{GenomeInfoDb}}.
Valid options include \code{"ncbi"}, \code{"ensembl"}, \code{"ucsc"} and
\code{"dbsnp"}. The following table provides a preview of each style (note
\code{ncbi} and \code{ensembl} are identical):

\tabular{rrr}{
  ncbi/ensembl  \tab ucsc  \tab dbsnp\cr
  1             \tab chr1  \tab ch1 \cr
  2             \tab chr2  \tab ch2\cr
  3             \tab chr3  \tab ch3\cr
  \dots         \tab \dots \tab \dots\cr
  X             \tab chrX  \tab chX\cr
  Y             \tab chrY  \tab chY\cr
  MT            \tab chrM  \tab chMT
}
}

\section{Preprocessing}{


The \code{preprocessor} argument allows you to specify shell commands that
preprocess the file before it's read into \R. For example, we could use
\code{grep} to filter our results to include only markers with an RS number:

\preformatted{ read_gwas("my-results.txt", preprocess = "grep -e '^rs'") }

Note that \code{read_gwas()} handles the header row separately so column
labels wouldn't be filtered out by \code{grep} in this example.

By default, the input filename is appended to \code{preprocess} argument
prior to execution. However, you can control where the filename should be
inserted in the command by using \code{\%s} as a placeholder. In the following
example, \code{tr} is being used to remove null terminators:

\preformatted{
  read_gwas("my-results.txt", preprocess = "tr -d '\\000' < \%s")
}
}

