% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.r
\name{set_chromosomes}
\alias{set_chromosomes}
\title{Convert chromosomes column to a propery ordered factor}
\usage{
set_chromosomes(x, chromosome_style)
}
\arguments{
\item{x}{vector of chromosome values}

\item{chromosome_style}{Convert chromosomes to ordered factors with labels
based on the specified style (default is \code{"ucsc"}; see below for a
comparison of the different styles). Set to \code{NULL} to leave
chromosomes unchanged.}
}
\description{
Convert chromosomes column to a propery ordered factor
}
